fix getConfirmBookingPageByMember, convertToTicketByMember, getConfir…

+++ MovieBooking 2/src/main/java/com/example/MovieBooking/service/impl/BookingServiceImpl.java
@@ -72,7 +72,7 @@ public class BookingServiceImpl implements IBookingService {
         LocalDate date = booking.getShowDate().getShowDate();
         String time = booking.getSchedule().getScheduleTime();
         int status = booking.getStatus();
-        int useScore = booking.getUseScore();
+        Long useScore = booking.getUseScore();
         String screen = booking.getMovie().getCinemaRoom().getScreen();
         long total = booking.getTotalMoney();
         StringBuilder seatString = new StringBuilder();
